Refactor : responseHandler split up
ResponseHandler
> KeywordResponseHandler
> ResponseHandler

해결: close
